All file utils not related to pipeline to a test file utils class
