+++ src/main/java/uk/ac/ebi/eva/utils/FileUtils.java
@@ -32,30 +32,6 @@ public abstract class FileUtils {
         return URLHelper.createUri(path);
     }
 
-    @Deprecated
-    public static File makeGzipFile(String content, String vepOutput) throws IOException {
-        File tempFile = new File(vepOutput);
-        try (FileOutputStream output = new FileOutputStream(tempFile)) {
-            try (Writer writer = new OutputStreamWriter(new GZIPOutputStream(output), "UTF-8")) {
-                writer.write(content);
-            }
-        }
-        return tempFile;
-    }
-
-    public static void copyResource(String resourcePath, String outputDir) throws IOException {
-        File vcfFile = new File(FileUtils.class.getResource(resourcePath).getFile());
-        FileCopyUtils.copy(vcfFile, new File(outputDir, resourcePath));
-    }
-
-    public static File getResource(String resourcePath) {
-        return new File(FileUtils.class.getResource(resourcePath).getFile());
-    }
-
-    public static URL getResourceUrl(String resourcePath) {
-        return FileUtils.class.getResource(resourcePath);
-    }
-
     public static Resource getResource(File file) throws IOException {
         Resource resource;
         if (CompressionHelper.isGzip(file)) {
