+++ src/test/java/uk/ac/ebi/eva/pipeline/jobs/steps/PedLoaderStepTest.java
@@ -39,7 +39,7 @@ import java.util.stream.Collectors;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
-import static uk.ac.ebi.eva.utils.FileUtils.getResource;
+import static uk.ac.ebi.eva.test.utils.TestFileUtils.getResource;
 
 /**
  * Test for {@link PedLoaderStep}
